+++ POSSystem2/app/src/main/java/com/example/posystem2/ItemModel.kt
@@ -4,10 +4,11 @@ data class ItemModel(
     val orderId: Int,
     val imageUri: String,
     val itemName: String,
-    val itemPrice: Int,
-    var quantity: Int = 1,  // New field to mark the quantity of the item
-    val ordered: Boolean = false  // Field to mark if the item is part of an order
+    val itemPrice: Float,
+    var quantity: Int = 1,
+    val ordered: Boolean = false
 )
+
